% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_FR.R
\docType{data}
\name{covid19_FR}
\alias{covid19_FR}
\title{Key numbers about covid in France}
\format{
A data frame with 2123 rows and 10 variables
}
\source{
opencovid-19 data \url{https://github.com/opencovid19-fr/data/blob/master/dist/chiffres-cles.csv}
stored in \code{'data/raw/chiffres-cles.csv'}
}
\usage{
covid19_FR
}
\description{
A dataset containing the community consolidated data on covid-19 epidemics in France
as of 2020-03-23 at 11:00 am
}
\examples{
data(covid19_FR)
#read.csv("https://github.com/opencovid19-fr/data/raw/master/dist/chiffres-cles.csv")

}
\keyword{datasets}
