% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/french_regions_map.R
\name{french_regions_map}
\alias{french_regions_map}
\title{Map of French region}
\usage{
french_regions_map(
  fill_info_df,
  mytitle,
  one_out_of = 30,
  show_labels = TRUE,
  col_labels = "white",
  size_labels = 5
)
}
\description{
Map of French region
}
