% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_covid19.R
\name{get_data_covid19}
\alias{get_data_covid19}
\title{Scrap and format data from opencovid19-fr}
\usage{
get_data_covid19(
  maille = "Nouvelle-Aquitaine",
  source = "ARS Nouvelle-Aquitaine",
  date_start = NULL,
  date_end = NULL
)
}
\description{
Scrap and format data from opencovid19-fr
}
\examples{
get_data_covid19(maille = "Grand-Est",
                source = "ARS Grand-Est")
}
