% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_covid19_cumul_bylevel.R
\name{get_data_covid19_cumul_bylevel}
\alias{get_data_covid19_cumul_bylevel}
\title{Scrap and format data from opencovid19-fr}
\usage{
get_data_covid19_cumul_bylevel(
  level_geo = c("region", "departement"),
  source_ch = "sante-publique-france-data",
  update_from_source_url = TRUE
)
}
\arguments{
\item{source_ch}{a character string either \code{"sante-publique-france"}
or \code{"agences-regionales-sante"} or \code{"ministere-sante"}.
Other sources are available but not so reliable...}

\item{maille_cd}{a character sting indicating the code}

\item{epidemic_start}{a logical indicating wehter to left-truncate data at the
start of the epidemics. The start of the epidemics is the first date}
}
\description{
Scrap and format data from opencovid19-fr
}
\examples{

temp <- get_data_covid19_cumul_bylevel(level_geo = "departement")
write.table(temp, file = "data/SantePubliqueFranceData_fromOpencovid19FR_20200503.tsv",
           sep="\t", row.names=FALSE)
}
