% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_covid19_hierar_bylevel.R
\name{get_data_covid19_hierar_bylevel}
\alias{get_data_covid19_hierar_bylevel}
\title{Get data from opencovid19-fr for all french region or dpt}
\usage{
get_data_covid19_hierar_bylevel(
  level = "region",
  sources_hierarchy = c("sante-publique-france", "ministere-sante",
    "agences-regionales-sante", "opencovid19-fr"),
  date_start = NULL,
  date_end = NULL,
  update_from_source = FALSE,
  epidemic_start = TRUE,
  metropole = TRUE
)
}
\arguments{
\item{maille_cd}{a character sting indicating the code}

\item{source3}{a character string either \code{"SPF"}
or \code{"ARS"}}
}
\description{
Get data from opencovid19-fr for all french region or dpt
}
\examples{
# regions
opencovid19_consolidated_reg <- get_data_covid19_hierar_bylevel()
#save(opencovid19_consolidated_reg, file = "data/opencovid19_consolidated_reg.RData")

# departements
get_data_covid19_bylevel(level = "departement", metropole=FALSE)


}
