% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_covid19_hierarchical.R
\name{get_data_covid19_hierarchical}
\alias{get_data_covid19_hierarchical}
\title{Scrap and format data from opencovid19-fr}
\usage{
get_data_covid19_hierarchical(maille_cd = "FRA",
  sources_hierarchy = c("sante-publique-france", "ministere-sante",
  "agences-regionales-sante", "opencovid19-fr"), date_start = NULL,
  date_end = NULL, metropole = FALSE, update_from_source_url = FALSE,
  epidemic_start = TRUE)
}
\arguments{
\item{maille_cd}{a character sting indicating the code}

\item{epidemic_start}{a logical indicating wehter to left-truncate data at the
start of the epidemics. The start of the epidemics is the first date}

\item{source_ch}{a character string either \code{"sante-publique-france"}
or \code{"agences-regionales-sante"} or \code{"ministere-sante"}.
Other sources are available but not so reliable...}
}
\description{
Scrap and format data from opencovid19-fr
}
\examples{

data_FR <- get_data_covid19_hierarchical(maille_cd = "FRA")
View(data_FR)

data_NA <- get_data_covid19_hierarchical(maille_cd = "REG-75")
View(data_NA)

}
