% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_sivic_incid.R
\name{get_data_sivic_incid}
\alias{get_data_sivic_incid}
\title{Scrap and format data from SI-VIC on data.gouv from Santé Publique France}
\usage{
get_data_sivic_incid(update_from_source_url = TRUE)
}
\arguments{
\item{level_geo}{a character sting indicating}

\item{source_ch}{a character string either \code{"sante-publique-france"}
or \code{"agences-regionales-sante"} or \code{"ministere-sante"}.
Other sources are available but not so reliable...}

\item{epidemic_start}{a logical indicating wehter to left-truncate data at the
start of the epidemics. The start of the epidemics is the first date}
}
\description{
Scrap and format data from SI-VIC on data.gouv from Santé Publique France
}
\examples{

sivic_incid_20200503 <- get_data_sivic_incid()
#save(sivic_incid_20200503, file = "data/sivic_incid_20200503.RData")
#write.table(sivic_incid_20200503, file = "data/SantePubliqueFranceData_fromDataGouv_20200503_incid.tsv",
#            sep="\t", row.names=FALSE)
}
