% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencovid19_FR.R
\docType{data}
\name{opencovid19_FR}
\alias{opencovid19_FR}
\title{Key numbers about covid in France}
\format{
A data frame with 2123 rows and 10 variables
}
\source{
opencovid-19 data \url{https://github.com/opencovid19-fr/data/blob/master/dist/chiffres-cles.csv}
stored in \code{'data/raw/chiffres-cles_25mars00h30.csv'}
}
\usage{
opencovid19_FR
}
\description{
A dataset containing the community consolidated data on covid-19 epidemics in France
as of 2020-03-25 at 00:30 am
}
\examples{
data(opencovid19_FR)
#read.csv("data/raw/chiffres-cles_25mars00h30.csv")
#opencovid19_FR <- read.csv("https://github.com/opencovid19-fr/data/raw/master/dist/chiffres-cles.csv")

}
\keyword{datasets}
