% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seidrah_estim.R
\name{plot.seidrah_estim}
\alias{plot.seidrah_estim}
\title{Plotting method for a SEIDRAH fit object}
\usage{
\method{plot}{seidrah_estim}(x, log = FALSE)
}
\description{
Plotting method for a SEIDRAH fit object
}
