% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.seirah_solve.R
\name{plot.seidrah_solve}
\alias{plot.seidrah_solve}
\title{Plotting results from SEIDRAH solve}
\usage{
\method{plot}{seidrah_solve}(x, chiffres = NULL, split = TRUE)
}
\arguments{
\item{x}{an object of class \code{seirah_solve}}
}
\description{
Plotting results from SEIDRAH solve
}
