% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.seirah_solve.R
\name{plot.seirah_solve}
\alias{plot.seirah_solve}
\title{Plotting results from SEIRAH solve}
\usage{
\method{plot}{seirah_solve}(x, chiffres = NULL, split = TRUE)
}
\arguments{
\item{x}{an object of class \code{seirah_solve}}
}
\description{
Plotting results from SEIRAH solve
}
