% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seirah_estim.R
\name{plot_list.seirah_estim}
\alias{plot_list.seirah_estim}
\title{Plotting method for a list of SEIRAH fit objects}
\usage{
plot_list.seirah_estim(
  x,
  only_observed_dates = TRUE,
  compartments = c("S", "E", "I", "R", "A", "H")
)
}
\description{
Plotting method for a list of SEIRAH fit objects
}
