% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popreg.R
\docType{data}
\name{popreg}
\alias{popreg}
\title{Population numbers in France}
\format{
A data frame with 19 rows and 3 variables:
\describe{
  \item{maille_nom}{name in French}
  \item{maille_code}{administrative code}
  \item{population}{population}
}
}
\source{
INED \url{https://www.ined.fr/fichier/s_rubrique/159/estim.pop.nreg.sexe.gca.1975.2020.fr.xls}
stored in \code{'data/raw/population_regionsFR_INED1erjan2020.tsv'}
}
\usage{
popreg
}
\description{
A dataset containing the populations of France as well as broken down by region
as of 2020-01-01 (according to INED)
}
\examples{
data(popreg)
# popreg <- read.delim(system.file("extdata", "population_regionsFR_INED1erjan2020.tsv", encoding = "UTF8"))

}
\keyword{datasets}
