% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.sierah_estim.R
\name{predict.seirah_estim}
\alias{predict.seirah_estim}
\title{Prediction of Epidemics and ICU utilization}
\usage{
\method{predict}{seirah_estim}(object, thresholdICU = 5000, verbose = TRUE)
}
\arguments{
\item{object}{an object of class \code{seirah_estim}}
}
\description{
Prediction of Epidemics and ICU utilization
}
\examples{
data_FRA <- get_data_covid19(maille_cd = "FRA",
                             source_ch = "sante-publique-france")
fit_FRA <- seirah_estim(binit = c(log(1.75), log(0.41)),
                        data = data_FRA, verbose = FALSE)
pred <- predict(fit_FRA, verbose=TRUE)

}
