% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.sierah_estim.R
\name{predict.sierah}
\alias{predict.sierah}
\title{Prediction of Epidemics and ICU utilization}
\usage{
\method{predict}{sierah}(
  x,
  b,
  newdata,
  threesholdICU,
  alpha = 1,
  De = 5.2,
  Di = 2.3,
  Dq = 10,
  Dh = 30,
  popSize = 6.5e+07,
  dailyMove = 1e+06,
  log.print = TRUE,
  plot.comparison = FALSE
)
}
\arguments{
\item{x}{an object of class \code{seirah_estim}}
}
\description{
Prediction of Epidemics and ICU utilization
}
\examples{
pred <- predict(c(0.1912957, -0.5439555),
data=dataFR,threesholdICU=5000)
pred

}
