% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seidrah_ols.R
\name{seidrah_ols}
\alias{seidrah_ols}
\title{OLS Estimation of parameters in SIEDRAH models}
\usage{
seidrah_ols(
  unknowns,
  stateinit,
  data,
  alpha = 1,
  De = 5.2,
  Di = 2.3,
  Dq = 10,
  Dh = 30,
  popSize = 6.5e+07,
  dailyMove = 1e+06,
  timeconf = 1000,
  lengthconf = 1000,
  newdailyMove = 1e-05 * popSize,
  factorreductrans = 3,
  verbose = TRUE,
  obs = c("S", "E", "I", "D", "R", "A", "H")
)
}
\description{
OLS Estimation of parameters in SIEDRAH models
}
\examples{


}
