% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seidrah_solve.R
\name{seidrah_solve}
\alias{seidrah_solve}
\title{SEIRAH solve function}
\usage{
seidrah_solve(init, t, par)
}
\description{
SEIRAH solve function
}
\examples{
# Set parameter values
b<-0
r<-1
alpha<-1
De<-5.2
Di<-2.3
Dq<-2
Dh<-30
popSize<-10000000
dailyMove<-0

init <- c(9999467, 346, 80, 0, 0, 80, 27)
t<-seq(0,365)
par <- c(1, 0.7, 0.01, alpha, De, Di, Dq, Dh, popSize, dailyMove, 30,30,0,3000000)

# Solve system using lsoda
sol <- seidrah_solve(init, t, par)
plot(sol)

# Plot solution
plot(sol$time,sol$I,ylim=c(0,100))
}
