% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seirah_estim.R
\name{seirah_estim}
\alias{seirah_estim}
\title{Fitting a SEIRAH by OLS}
\usage{
seirah_estim(
  binit,
  data = NULL,
  stateinit = NULL,
  initwithdata = TRUE,
  alpha = 1,
  De = 5.2,
  Di = 2.3,
  Dq = 10,
  Dh = 30,
  popSize = 6.5e+07,
  dailyMove = 0.1 * popSize,
  verbose = TRUE,
  optim_ols = TRUE
)
}
\description{
Fitting a SEIRAH by OLS
}
\examples{
data_FRA <- get_data_covid19(maille_cd = "FRA",
                             source_ch = "sante-publique-france")
fit_FRA <- seirah_estim(binit = c(1.75, 0.41),
                        data = data_FRA)
plot(fit_FRA)

data_GE <- get_data_covid19(maille_cd = "REG-44",
                            source_ch = "agences-regionales-sante")
fit_GE <- seirah_estim(binit = c(1.75, 0.41),
                       data=data_GE, popSize = 5518000, dailyMove = 0.1*5518000)
plot(fit_GE)

data_NA <- get_data_covid19(maille_cd = "REG-75",
                           source_ch = "agences-regionales-sante")
fit_NA <- seirah_estim(binit = c(1.75, 0.41),
                      data = data_NA,
                      popSize = 5987000, dailyMove = 0.1*5987000)
plot(fit_NA)


simul_xihong<- seirah_estim(binit=c(1.75, 0.41),data=NULL,stateinit=c(9999467,346,80,0,80,27),
                            initwithdata=FALSE,alpha=1,De=5.2,Di=2.3,Dq=10,Dh=30,popSize=10000000,dailyMove=500000,verbose = TRUE,optim_ols = FALSE)
plot_list.seirah_estim(list(simul_xihong))
}
