% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sierah_ols.R
\name{seirah_ols}
\alias{seirah_ols}
\title{OLS Estimation of parameters in SIERAH models}
\usage{
seirah_ols(
  b,
  data,
  alpha = 1,
  De = 5.2,
  Di = 2.3,
  Dq = 10,
  Dh = 30,
  popSize = 6.5e+07,
  dailyMove = 1e+06,
  verbose = TRUE,
  fit = FALSE
)
}
\description{
OLS Estimation of parameters in SIERAH models
}
\examples{

data_FRA <- get_data_covid19(maille_cd = "FRA",
                           source_ch = "sante-publique-france")
#Perform OLS
optimresult_FRA <- optimx(c(log(1.75), log(0.41)),
                      fn = seirah_ols, data=data_FRA,
                      method = "Nelder-Mead")
#Plot fitted vs. observed
fit_FRA <- seirah_ols(b = c(optimresult_FRA$p1, optimresult_FRA$p2),
                  data = data_FRA, fit = TRUE)
plot(fit_FRA)


data_GE <- get_data_covid19(maille_cd = "REG-44",
                           source_ch = "agences-regionales-sante")
#optimresult_GE <- optim(par = c(log(1.75),log(0.41)), fn = seirah_ols,
                       data = data_GE, popSize=5518000 ,dailyMove=0.1*5518000)
#plot(seirah_ols(optimresult_GE$par, data=data_GE, popSize=5518000,dailyMove=0.1*5518000, fit = TRUE))
optimxresult_GE <- optimx(c(log(1.75),log(0.41)),
                     fn = seirah_ols, data=data_GE, popSize=5518000 ,dailyMove=0.1*5518000,
                     method = "Nelder-Mead")
fit_GE <- seirah_ols(b = c(optimxresult_GE$p1, optimxresult_GE$p2),
                  data=data_GE, popSize=5518000 ,dailyMove=0.1*5518000, fit = TRUE)
plot(fit_GE)




data_NA <- get_data_covid19(maille_cd = "REG-75",
                           source_ch = "agences-regionales-sante")
data_NA <- data_NA[-c(1:40), ]
data_NA$day <- data_NA$day-40

#optimresult_NA<-optim(c(log(1.75),log(0.41)), seirah_ols, data=data_NA, popSize=5987000 ,dailyMove=0.1*5987000)
#plot(seirah_ols(optimresult$par,data=data_NA, fit = TRUE, popSize=5987000 ,dailyMove=0.1*5987000))

optimxresult_NA <- optimx(c(log(1.75), log(0.41)),
                        fn = seirah_ols, data=data_NA, popSize=5987000 ,dailyMove=0.1*5987000,
                        method = "Nelder-Mead")
#Plot fitted vs. observed
fit_FRA <- seirah_ols(b = c(optimxresult_NA$p1, optimxresult_NA$p2),
                      data = data_NA, popSize=5987000 ,dailyMove=0.1*5987000, fit=TRUE)
plot(fit_FRA)

}
