% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seirah_sens_param_ode.R
\name{seirah_sens_param_ode}
\alias{seirah_sens_param_ode}
\title{SEIRAH derivative functions}
\usage{
seirah_sens_param_ode(t, X, par)
}
\description{
SEIRAH derivative functions
}
\keyword{internal}
