% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seirah_solve.R
\name{seirah_solve}
\alias{seirah_solve}
\title{SEIRAH solve function}
\usage{
seirah_solve(init, t, par)
}
\description{
SEIRAH solve function
}
\examples{
# Set parameter values
b<-1
r<-1
alpha<-1
De<-5.2
Di<-2.3
Dq<-2
Dh<-30
popSize<-10000000
dailyMove<-0

init<-c(9999467, 346, 80, 0, 80, 27)
t<-seq(0,365)
par <- c(b, r, alpha, De, Di, Dq, Dh, popSize, dailyMove)

# Solve system using lsoda
sol <- seirah_solve(init, t, par)

# Plot solution
plot(sol)
}
