% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sierah_ols.R
\name{sierah_ols}
\alias{sierah_ols}
\title{OLS Estimation of parameters in SIERAH models}
\usage{
sierah_ols(
  b,
  data,
  alpha = 1,
  De = 5.2,
  Di = 2.3,
  Dq = 10,
  Dh = 30,
  popSize = 6.5e+07,
  dailyMove = 1e+06,
  verbose = TRUE,
  plot_comparison = FALSE
)
}
\description{
OLS Estimation of parameters in SIERAH models
}
\examples{
dataFR<-get_data_covid19(maille = "France",
             source = "Santé publique France")
#Perform OLS            
optimresult<-optim(c(log(1.75),log(0.41)), ols_sierah,data=dataFR)
#Plot fitted vs. observed 
ols_sierah(optimresult$par,data=dataFR,plot.comparison = TRUE)


GrandEST<-get_data_covid19(maille = "Grand-Est",
              source = "ARS Grand-Est")
optimresult<-optim(c(log(1.75),log(0.41)), ols_sierah,data=GrandEST,popSize=5518000 ,dailyMove=0.1*5518000)
ols_sierah(optimresult$par,data=GrandEST,popSize=5518000 ,dailyMove=0.1*5518000,plot.comparison = TRUE)


NouvelleAquitaine<-get_data_covid19(maille = "Nouvelle-Aquitaine",
              source = "ARS Nouvelle-Aquitaine")
NouvelleAquitaine<-NouvelleAquitaine[41:54,]
NouvelleAquitaine$day<-NouvelleAquitaine$day-40
optimresult<-optim(c(log(1.75),log(0.41)), ols_sierah,data=NouvelleAquitaine,popSize=5987000 ,dailyMove=0.1*5987000)
ols_sierah(optimresult$par,data=NouvelleAquitaine,plot.comparison = TRUE,popSize=5987000 ,dailyMove=0.1*5987000)
}
