% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seirah_solve_der_X_theta.R
\docType{data}
\name{sol_X_dX_dtheta}
\alias{sol_X_dX_dtheta}
\title{# Solve system using lsoda}
\format{
An object of class \code{deSolve} (inherits from \code{matrix}) with 11 rows and 49 columns.
}
\usage{
sol_X_dX_dtheta
}
\description{
# Solve system using lsoda
}
\keyword{datasets}
