% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_covid19.R
\name{get_data_covid19}
\alias{get_data_covid19}
\title{Scrap and format data from opencovid19-fr}
\usage{
get_data_covid19(
  maille_cd = "FRA",
  source_ch = "sante-publique-france",
  date_start = NULL,
  date_end = NULL,
  metropole = FALSE,
  update_from_source_url = FALSE,
  epidemic_start = TRUE
)
}
\arguments{
\item{maille_cd}{a character sting indicating the code}

\item{source_ch}{a character string either \code{"sante-publique-france"}
or \code{"agences-regionales-sante"}.
Other sources are available but not so reliable...}
}
\description{
Scrap and format data from opencovid19-fr
}
\examples{

get_data_covid19(maille_cd = "FRA",
                source_ch = "ministere-sante")

get_data_covid19(maille_cd = "DPT-33",
                source_ch = "agences-regionales-sante")

get_data_covid19(maille_cd = "REG-75",
                source_ch = "agences-regionales-sante")

get_data_covid19(maille_cd = "REG-44",
                source_ch = "agences-regionales-sante")

get_data_covid19(maille_cd = "FRA",
                source_ch = "sante-publique-france")

get_data_covid19(maille_cd = "WORLD",
                source_ch = "sante-publique-france")
}
