% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_covid19_bylevel.R
\name{get_data_covid19_bylevel}
\alias{get_data_covid19_bylevel}
\title{Get data from opencovid19-fr for all french region or dpt}
\usage{
get_data_covid19_bylevel(
  level = "region",
  source3 = "SPF",
  date_start = NULL,
  date_end = NULL,
  update_from_source = FALSE
)
}
\arguments{
\item{maille_cd}{a character sting indicating the code}

\item{source_ch}{a character string either \code{"Santé publique France"}
or \code{"ARS XXX-XXX"} (where 'XXX-XXX' must be replaced by the region name).
Other sources are available but not so reliable...}
}
\description{
Get data from opencovid19-fr for all french region or dpt
}
\examples{
get_data_covid19(maille_cd = "DPT-33",
                source_ch = "ARS Nouvelle-Aquitaine")

get_data_covid19(maille_cd = "REG-75",
                source_ch = "ARS Nouvelle-Aquitaine")

get_data_covid19(maille_cd = "REG-44",
                source_ch = "ARS Grand-Est")

get_data_covid19(maille_cd = "FRA",
                source_ch = "Santé publique France")

get_data_covid19(maille_cd = "WORLD",
                source_ch = "Santé publique France")
}
