% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seirah_ols.R
\name{seirah_ols}
\alias{seirah_ols}
\title{OLS Estimation of parameters in SIERAH models}
\usage{
seirah_ols(
  b,
  stateinit,
  data,
  alpha = 1,
  De = 5.2,
  Di = 2.3,
  Dq = 10,
  Dh = 30,
  popSize = 6.5e+07,
  dailyMove = 1e+06,
  verbose = TRUE
)
}
\description{
OLS Estimation of parameters in SIERAH models
}
