% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sierah_ols.R
\name{sierah_ols}
\alias{sierah_ols}
\title{OLS Estimation of parameters in SIERAH models}
\usage{
sierah_ols(
  b,
  data,
  alpha = 1,
  De = 5.2,
  Di = 2.3,
  Dq = 10,
  Dh = 30,
  popSize = 6.5e+07,
  dailyMove = 1e+06,
  verbose = TRUE,
  plot_comparison = FALSE
)
}
\description{
OLS Estimation of parameters in SIERAH models
}
\examples{
dataFR<-get_data_covid19(maille = "France",
             source = "Santé publique France")
#Perform OLS
optimresult<-optim(c(log(1.75),log(0.41)), sierah_ols,data=dataFR)
#Plot fitted vs. observed
sierah_ols(optimresult$par,data=dataFR,plot_comparison = TRUE)


GrandEST<-get_data_covid19(maille = "Grand-Est",
              source = "ARS Grand-Est")
optimresult<-optim(c(log(1.75),log(0.41)), sierah_ols,data=GrandEST)
sierah_ols(optimresult$par,data=GrandEST,plot_comparison = TRUE)


NouvelleAquitaine<-get_data_covid19(maille = "Nouvelle-Aquitaine",
              source = "ARS Nouvelle-Aquitaine")
NouvelleAquitaine<-NouvelleAquitaine[41:54,]
optimresult<-optim(c(log(1.75),log(0.41)), sierah_ols,data=NouvelleAquitaine)
sierah_ols(optimresult$par,data=NouvelleAquitaine,plot_comparison = TRUE)
}
